requirements.txt
altair==5.3.0
beautifulsoup4==4.12.3
fake_useragent==1.5.1
google-cloud-bigquery-storage==2.25.0
groq==0.9.0
numpy==2.0.1
openai==1.37.0
pandas==2.2.2
pandas_gbq==0.23.1
plotly==5.9.0
protobuf==5.27.2
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.2.2
streamlit==1.36.0
tiktoken==0.7.0
tqdm==4.65.0
